Move JDA stuff to seperate class
