+++ src/main/java/be/maxim/mc_discord_bot/DiscordClient.java
@@ -3,18 +3,15 @@ package be.maxim.mc_discord_bot;
 import net.dv8tion.jda.api.OnlineStatus;
 import net.dv8tion.jda.api.sharding.DefaultShardManagerBuilder;
 import net.dv8tion.jda.api.sharding.ShardManager;
-import org.bukkit.plugin.java.JavaPlugin;
-
+import java.io.File;
+import java.util.Scanner;
 import javax.security.auth.login.LoginException;
-import java.util.Objects;
-//import java.io.File;
-//import java.util.Scanner;
 
-public final class Mc_discord_bot extends JavaPlugin {
+public class DiscordClient {
 
     private final ShardManager shardManager;
 
-    public Mc_discord_bot() throws LoginException {
+    public DiscordClient() throws LoginException {
         DefaultShardManagerBuilder builder = DefaultShardManagerBuilder.createDefault("TOKEN");
         builder.setStatus(OnlineStatus.ONLINE);
         shardManager = builder.build();
@@ -24,20 +21,12 @@ public final class Mc_discord_bot extends JavaPlugin {
         return shardManager;
     }
 
-    @Override
-    public void onEnable() {
-        System.out.println("Mc_discord_bot is starting...");
-        Objects.requireNonNull(this.getCommand("test")).setExecutor(new CommandTest());
-
+    public void StartBot(){
         try {
-            Mc_discord_bot bot = new Mc_discord_bot();
+            DiscordClient bot = new DiscordClient();
         } catch (LoginException e) {
             System.out.println("ERROR: Discord ogin exception");
         }
     }
 
-    @Override
-    public void onDisable() {
-        System.out.println("Mc_discord_bot is stopping...");
-    }
 }
