+++ src/main/resources/plugin.yml
@@ -1,6 +1,6 @@
 name: mc_discord_bot
 version: '1.0-SNAPSHOT'
-main: be.maxim.mc_discord_bot.Mc_discord_bot
+main: be.maxim.mc_discord_bot.McDiscordBot
 api-version: '1.21'
 commands:
   test:
